Major changes to project structure and clear code and remove file uploding method
