+++ src/validaters/validateVehicleData.js
